+++ ci_tools/bot_tools/bot_funcs.py
@@ -14,6 +14,7 @@ default_python_versions = {
         'anaconda_windows': '3.10',
         'coverage': '3.8',
         'docs': '3.8',
+        'intel': '3.9',
         'linux': '3.8',
         'macosx': '3.11',
         'pickle_wheel': '3.8',
@@ -30,6 +31,7 @@ test_names = {
         'anaconda_windows': "Unit tests on Windows with anaconda",
         'coverage': "Coverage verification",
         'docs': "Check documentation",
+        'intel': "Unit tests on Linux with Intel compiler",
         'linux': "Unit tests on Linux",
         'macosx': "Unit tests on MacOSX",
         'pickle_wheel': "Test pickling during wheel installation",
@@ -415,7 +417,7 @@ class Bot:
             True if the test should be run, False otherwise.
         """
         print("Checking : ", name)
-        if key in ('linux', 'windows', 'macosx', 'anaconda_linux', 'anaconda_windows', 'coverage'):
+        if key in ('linux', 'windows', 'macosx', 'anaconda_linux', 'anaconda_windows', 'coverage', 'intel'):
             has_relevant_change = lambda diff: any((f.startswith('pyccel/') or f.startswith('tests/')) \
                                                     and f.endswith('.py') and f != 'pyccel/version.py' \
                                                     for f in diff) #pylint: disable=unnecessary-lambda-assignment
@@ -428,7 +430,7 @@ class Bot:
             has_relevant_change = lambda diff: any(f.endswith('.py') and f != 'pyccel/version.py' \
                                                     for f in diff) #pylint: disable=unnecessary-lambda-assignment
         elif key in ('spelling'):
-            has_relevant_change = lambda diff: any(f.endswith('.md') for f in diff) #pylint: disable=unnecessary-lambda-assignment
+            has_relevant_change = lambda diff: any(f.endswith('.md') or f == '.dict_custom.txt' for f in diff) #pylint: disable=unnecessary-lambda-assignment
         elif key in ('pickle', 'pickle_wheel', 'editable_pickle'):
             has_relevant_change = lambda diff: any(f.startswith('pyccel/') and f.endswith('.py') \
                                                     and f != 'pyccel/version.py' for f in diff) #pylint: disable=unnecessary-lambda-assignment
@@ -523,7 +525,13 @@ class Bot:
             self.mark_as_draft()
             return False
 
-        welcome_comment = next(c for c in self._GAI.get_comments(self._pr_id) if c['user']['type'] == 'Bot' and c['body'].startswith('Hello'))
+        try:
+            welcome_comment = next(c for c in self._GAI.get_comments(self._pr_id) if c['body'].startswith('Here is your checklist.'))
+        except StopIteration:
+            self._GAI.create_comment(self._pr_id, message_from_file('missing_checklist.txt'))
+            self.mark_as_draft()
+            return False
+
         if '- [ ]' in welcome_comment['body']:
             self._GAI.create_comment(self._pr_id, message_from_file('set_draft_checklist_incomplete.txt').format(url = welcome_comment['html_url']))
             self.mark_as_draft()
@@ -654,8 +662,8 @@ class Bot:
         """
         print("Trusted?")
         in_team = self._GAI.check_for_user_in_team(user, 'pyccel-dev')
-        if in_team["message"] != "Not found":
-            print("In team")
+        if "message" not in in_team:
+            print("In team: ", in_team)
             return True
         print("User not in team")
         merged_prs = self._GAI.get_prs('all')
@@ -896,6 +904,27 @@ class Bot:
                                  reply_to = comment_id)
         print(reply.text)
 
+    def fill_checklist(self, comment_url, user):
+        """
+        Create the PR checklist for the user.
+
+        Create the PR checklist for the user.
+
+        Parameters
+        ----------
+        comment_url : str
+            The url where the comment was left.
+
+        user : str
+            The username of the person who left the comment.
+        """
+        body = message_from_file('checklist.txt')
+        merged_prs = self._GAI.get_prs('all')
+        has_merged_pr = any(pr for pr in merged_prs if pr['user']['login'] == user and pr['merged_at'])
+        if not has_merged_pr:
+            body += message_from_file('first_checklist.txt')
+        self._GAI.modify_comment(comment_url, body)
+
     def is_pr_draft(self):
         """
         Indicate whether the pull request is a draft.
@@ -908,6 +937,19 @@ class Bot:
             True if draft, False otherwise.
         """
         return self._pr_details['draft']
+
+    def is_pr_fork(self):
+        """
+        Indicate whether the pull request is created from a fork.
+
+        Indicate whether the pull request is created from a fork.
+
+        Returns
+        -------
+        bool
+            True if fork, False otherwise.
+        """
+        return self._pr_details['head']['repo']['full_name'] != 'pyccel/pyccel'
 
     def leave_comment(self, comment):
         """