+++ ci_tools/devel_branch_tests.py
@@ -14,3 +14,4 @@ if __name__ == '__main__':
     bot.run_tests(['pickle_wheel'], '3.8', force_run = True)
     bot.run_tests(['anaconda_linux'], '3.10', force_run = True)
     bot.run_tests(['anaconda_windows'], '3.10', force_run = True)
+    bot.run_tests(['intel'], '3.9', force_run = True)